Refactored typescript to allow multiple tables on a single page
 * moved calculator code into new class in calculator.ts
 * replaced ids with class names or with prefixed id
 * moved calculator instantiation out of typescript modules
