+++ metalation-calculator-wp/include/calculator.html
@@ -1,4 +1,4 @@
-<div class="metalation-calculator" id="metalation-calculator">
+<div class="metalation-calculator">
   <p class="metalation-calculator-intro">
     To predict the metalation state of a protein or molecule, fill in values in the table for
     as many determined metal affinities (and availabilities if known) as possible.
@@ -13,7 +13,7 @@
     </div>
   </div>
 
-<table id="metalation-table">
+<table class="metalation-table">
   <thead>
     <tr>
       <td></td>
@@ -33,11 +33,11 @@
     <td></td>
     <td></td>
     <td></td>
-    <th id="total-metalation" class="result"></th>
+    <th class="result total-metalation"></th>
   </tfoot>
 </table>
-<input type="button" disabled value="Download Results" id="download-btn"/>
-<input type="button" value="Reset" id="reset-btn"/>
+<input type="button" disabled value="Download Results" class="download-btn"/>
+<input type="button" value="Reset" class="reset-btn"/>
 
 <p class="footnote">The calculator is based on <a href="https://doi.org/10.1038/s41467-021-21479-8">Nature Communications 12 1195- (2021)</a>.
 Supported by grant <strong>BB/V006002/1</strong> from the Biotechnology and Biological Sciences Research Council.</p>
