+++ metalation-calculator-wp/include/main.css
@@ -1,4 +1,4 @@
-.metalation-calculator #metalation-table {
+.metalation-calculator .metalation-table {
   table-layout: fixed;
   width: 100%;
   border-collapse: separate;
@@ -7,56 +7,56 @@
   font-size: 16px;
   /* Hide number up/down arrows in Chrome, Safari, Edge, Opera */
 }
-.metalation-calculator #metalation-table td,
-.metalation-calculator #metalation-table th {
+.metalation-calculator .metalation-table td,
+.metalation-calculator .metalation-table th {
   padding: 2px 5px;
   border: 0;
   justify-content: center;
 }
-.metalation-calculator #metalation-table th {
+.metalation-calculator .metalation-table th {
   text-align: center;
 }
-.metalation-calculator #metalation-table td,
-.metalation-calculator #metalation-table .result {
+.metalation-calculator .metalation-table td,
+.metalation-calculator .metalation-table .result {
   text-align: right;
   padding-right: 20px;
 }
-.metalation-calculator #metalation-table thead,
-.metalation-calculator #metalation-table tfoot {
+.metalation-calculator .metalation-table thead,
+.metalation-calculator .metalation-table tfoot {
   background-color: #88cede;
   padding: 1.2%;
 }
-.metalation-calculator #metalation-table sup {
+.metalation-calculator .metalation-table sup {
   white-space: nowrap;
 }
-.metalation-calculator #metalation-table tbody .grouped {
+.metalation-calculator .metalation-table tbody .grouped {
   background-color: rgba(96, 96, 96, 0.2);
 }
-.metalation-calculator #metalation-table tbody th {
+.metalation-calculator .metalation-table tbody th {
   text-align: left;
   justify-content: center;
 }
-.metalation-calculator #metalation-table tbody th .metal-symbol {
+.metalation-calculator .metalation-table tbody th .metal-symbol {
   display: inline-block;
   margin-left: 15px;
 }
-.metalation-calculator #metalation-table tbody tr:nth-child(odd) {
+.metalation-calculator .metalation-table tbody tr:nth-child(odd) {
   background-color: white;
 }
-.metalation-calculator #metalation-table tbody tr:nth-child(odd) td.right-spacing {
+.metalation-calculator .metalation-table tbody tr:nth-child(odd) td.right-spacing {
   border-right: 15px solid white;
 }
-.metalation-calculator #metalation-table tbody tr:nth-child(even) {
+.metalation-calculator .metalation-table tbody tr:nth-child(even) {
   background-color: #ededed;
 }
-.metalation-calculator #metalation-table tbody tr:nth-child(even) td.right-spacing {
+.metalation-calculator .metalation-table tbody tr:nth-child(even) td.right-spacing {
   border-right: 15px solid #ededed;
 }
-.metalation-calculator #metalation-table .tooltip {
+.metalation-calculator .metalation-table .tooltip {
   position: relative;
   padding: 1.115%;
 }
-.metalation-calculator #metalation-table .tooltip .tooltiptext {
+.metalation-calculator .metalation-table .tooltip .tooltiptext {
   visibility: hidden;
   width: 120px;
   background-color: #606060;
@@ -69,7 +69,7 @@
   display: block;
   bottom: 100%;
 }
-.metalation-calculator #metalation-table .tooltip .tooltiptext::after {
+.metalation-calculator .metalation-table .tooltip .tooltiptext::after {
   content: " ";
   position: absolute;
   top: 100%;
@@ -80,13 +80,13 @@
   border-style: solid;
   border-color: #606060 transparent transparent transparent;
 }
-.metalation-calculator #metalation-table .tooltip:hover .tooltiptext {
+.metalation-calculator .metalation-table .tooltip:hover .tooltiptext {
   visibility: visible;
 }
-.metalation-calculator #metalation-table #total-metalation {
+.metalation-calculator .metalation-table #total-metalation {
   padding: 18px;
 }
-.metalation-calculator #metalation-table input[type=number] {
+.metalation-calculator .metalation-table input[type=number] {
   text-align: right;
   width: 100%;
   border-width: 1px;
@@ -97,18 +97,18 @@
   -moz-appearance: textfield;
   /* stylelint-enable */
 }
-.metalation-calculator #metalation-table input[type=number]::-webkit-outer-spin-button,
-.metalation-calculator #metalation-table input[type=number]::-webkit-inner-spin-button {
+.metalation-calculator .metalation-table input[type=number]::-webkit-outer-spin-button,
+.metalation-calculator .metalation-table input[type=number]::-webkit-inner-spin-button {
   /* stylelint-disable */
   -webkit-appearance: none;
   /* stylelint-enable */
   margin: 0;
 }
-.metalation-calculator #metalation-table .error-msg {
+.metalation-calculator .metalation-table .error-msg {
   color: red;
   font-size: smaller;
 }
-.metalation-calculator #metalation-table label {
+.metalation-calculator .metalation-table label {
   cursor: pointer;
   text-indent: -9999px;
   width: 30px;
@@ -119,7 +119,7 @@
   position: relative;
   top: 3px;
 }
-.metalation-calculator #metalation-table label::after {
+.metalation-calculator .metalation-table label::after {
   content: "";
   position: absolute;
   top: 2px;
@@ -131,15 +131,15 @@
   border-radius: 5px;
   transition: 0.5s;
 }
-.metalation-calculator #metalation-table input[type=checkbox] {
+.metalation-calculator .metalation-table input[type=checkbox] {
   height: 0;
   width: 0;
   visibility: hidden;
 }
-.metalation-calculator #metalation-table input[type=checkbox]:checked + label {
+.metalation-calculator .metalation-table input[type=checkbox]:checked + label {
   background: #bebebe;
 }
-.metalation-calculator #metalation-table input[type=checkbox]:checked + label::after {
+.metalation-calculator .metalation-table input[type=checkbox]:checked + label::after {
   left: 14px;
 }
 .metalation-calculator .metalation-table-header {
@@ -161,8 +161,8 @@
   max-width: 120px;
   margin-right: 10px;
 }
-.metalation-calculator #download-btn,
-.metalation-calculator #reset-btn {
+.metalation-calculator .download-btn,
+.metalation-calculator .reset-btn {
   display: inline;
   float: right;
   margin-right: 2px;
