+++ typescript/src/index.html
@@ -34,7 +34,15 @@
                     <h3>Metalation Calculator</h3>
                     <h4>Predict metal occupancies in vivo</h4>
 
-                    @@include('./includes/calculator.html')
+                    <div id="metalation-calculator">
+                      @@include('./includes/calculator.html')
+                    </div>
+
+                    <script type="text/javascript">
+                      window.addEventListener("DOMContentLoaded", () => {
+                        window.setupCalculator("metalation-calculator", "Idealised <em>Salmonella</em>");
+                      });
+                    </script>
                   </div>
                 </div>
                 <div class="page-standard__sidebar c-sidebar c-block--layout-right">
